requirements/app.txt
galaxy.plugin.api==0.69

# steam_network backend
protobuf==4.24.4
vdf==3.4
pyobjc-framework-CoreServices==9.2; sys_platform == 'darwin'
websockets==11.0.3
dataclasses-json==0.6.7
rsa==4.9.1
certifi